requirements.txt
alembic==1.4.2
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-Migrate==2.5.3
flask-mongoengine==0.9.5
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mongoengine==0.20.0
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-engineio==3.13.2
python-socketio==4.6.0
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
WTForms==2.3.3
